extract tarballs not generated with s3tar
added the ability to generate a toc for an existing tarball. The previous archive must not have compression.
for now the tar must be in the local system to generate the toc.

command:
s3tar --generate-toc -f my-previous-archive.tar -C /home/user/my-previous-archive.toc.csv

Then upload the tar and the toc to Amazon S3.
Once the tar is in Amazon S3 you can extract it by passing the --external-toc flag with the location:

s3tar --region us-west-2 \
    --external-toc s3://bucket/prefix/my-previous-archive.toc.csv \
    -xvf s3://bucket/prefix/my-previous-archive.tar \
    -C s3://bucket/prefix-dest/

This release introduces a breaking change in List, pin your versions if using the Go code.