in-memory tar
This update introduces in-memory tarball generation. The goal is to optimize s3tar's performance for small objects. The new `--concat-in-memory` flag optimizes the process, altering the core design to **download** and build MultiPart objects in-memory when enabled. This modification brings about a notable improvement in both performance and pricing, prioritizing GET requests over PUT requests. Importantly, the default behavior of s3tar remains unaffected to ensure continuity with existing workflows. Users that are creating tarballs of extensive small files, numbering in the hundreds of thousands or millions, are recommended to leverage the --concat-in-memory flag for enhanced efficiency.

- added the functionality to build every part of the MPU in-memory. To use this
  feature the --concat-in-memory flag needs to be passed.
- now that we can tar in memory, we can tar smaller files than the MPU limit
  and upload it as a single part
- concat-in-memory functionality has toc disabled
- added functionality to generate a toc remotely