Allow different destination region
To allow packing data from one region to an archive in different region, additional argument --dstRegion (and --dstEndpointUrl) are introduced and separate client is created if needed.

Change is backwards compatible, the client defaults to the same as the source one if neither dstRegion or dstEndpointUrl are specified.

In the code, I tried to disambiguate the source (without prefix) service/client from that is used for listing and getting object headers from the one used for putting object (copies) which can take the source object from different region easily.